Added new guns
Adds:
Disabler Magnum
a 3 shot stun with less charge in the gun. Rewards careful aim for the cunning captain

Viperfang.
A mean Assault rifle similar to the lecter, Uses .20 Rifle with 6 firerate. Not a bad choice if you want more than a lecter.

Overclocked Railgun,
3 shots. its rounds go SO fast they look like lasers! and burn through you!.
VERY slow self recharges. With 35 Heat damage.

Ceremonial Rifle.
Similar to the Mosin with a smaller clip and a faster firerate. and made of modern materials. This thing will do the trick nicely for your larp session.